requirements.txt
sounddevice
numpy
paho-mqtt